+++ aws/sqs/variables.tf
@@ -68,3 +68,15 @@ variable "principal_roles" {
   type        = list(string)
   default     = null
 }
+
+variable "read_policy_name" {
+  description = "The name of the custom read policy"
+  type        = string
+  default     = null
+}
+
+variable "write_policy_name" {
+  description = "The name of the custom write policy"
+  type        = string
+  default     = null
+}